Draw particles with transparency
We draw particles behind all other objects (planets, ships, bullets), because they faded to black over time: When on a black background (space), this looks perfectly fine, but them fading to black over a colored background (e.g. a yellow star) would look weird. They should fade to the background's color instead, which this achieves.

As a consequence, we can also draw particles on top of all other objects.

The exact solution used in the code here doesn't feel very idiomatic. It'd be more idiomatic to mix the colors after applying alpha to the particle's color, but pygame doesn't offer that.