Introduce abstraction between keys and controls
This is nothing novel, in fact it already exists: For the player, "Am I currently activating my left-rotation-thruster?" is not equivalent to "Am I currently pressing my "left-rotation-thruster" button?", because of `do_rotation_for_player`.

This refactoring doesn't change anything yet. It's in preparation for putting the `do_rotation_for_player` into the class that's now called "ShipInputTank".

After that's done and functional, we can start creating a new class called "ShipInputAbsolute", and maybe "ShipInputMouse".